Problem: omni_httpd does not support WebSockets
This is a major problem for anything remotely interesting these days.

Solution: devise the first version of WebSocket handling

Current approach definitely has deficiencies. For example, using UNIX
domain sockets can be problematic here. It's good that sending a message to
UNIX domain socket is directly connecting us with the consumer (the
networking thread) thus ensuring that we are not simply buffering the
messages. But if, for some reason, the consumer is not reading, it'll
block us.

An even bigger problem is that we currently won't try to send messages
to any other worker and will simply schedule messages for the current
buffer. It's okay for the initial prototype, but won't scale well.