Problem: PostgREST is not only used for relations
Some of its users heavily rely on its RPC capabilities, and this feature is critical to them.

Solution: introduce RPC API compatibility in omni_rest

It's an early draft for basic functionality. Not everything is implemented – most notably, there's no horizontal filtering
and not all of the function shapes are handled. Not because of any limitations but rather to make the basics ready and iterate later.