Problem: difficulty maintaining complex omni_sqlite code
The code is dealing with intricacies of Postges and SQLite and it is
sometimes difficult to tell what to do.

There are also some known bugs I'd like to fix.

Solution: rewrite it in C++ using cppgres

Thanks to @michelp (Michel Pelletier) for all the help explaining
expandable datums that went into cppgres!